Refactor on where the approximation for power convergence is placed. I want it to be in utils
